+++ tests/test_1q_QCut.py
@@ -49,50 +49,13 @@ def test_separate_subcircuits() -> None:
         cut_circuit = wc.get_locations_and_subcircuits(circ.copy())
         circs = cut_circuit.subcircuits
         _remove_obsm(circs)
-        print(count)
         count += 1
+
+        assert len(circs) == sq.number_of_subcircuits[solution_index]
+
         for circ_index, subcirc in enumerate(circs):
-            for op1, op2 in zip(
-                subcirc.data, sq.subcircuit_solutions[solution_index][circ_index]
-            ):
-                assert op1.operation.name == op2.operation.name  # noqa: S101
 
-
-# def test_get_experiment_circuits() -> None:
-#     """Test get_experiment_circuits function.
-
-#     This function tests whether the get_experiment_circuits method correctly
-#     generates the experiment circuits from the subcircuits and qubit subsets for
-#     each test circuit by comparing the operations in the generated experiment
-#     circuits to the pre-defined solutions.
-#     """
-#     for solution_index, circ in enumerate(sq.test_circuits):
-#         # Retrieve qubit locations and subcircuits
-#         qss, circuits = wc.get_locations_and_subcircuits(circ)
-
-#         # Generate experiment circuits
-#         experiment_circuits, coefs, id_meas = ck.get_experiment_circuits(circuits, 
-#                                                                           qss)
-
-#         assert np.array_equal(coefs, s.coefs_solutions[solution_index])
-
-#         assert np.array_equal(id_meas, s.id_meas_solutions[solution_index])
-#         # Flatten the nested list of experiment circuits and extract their data
-#         test_data = [
-#             op.data for sublist in experiment_circuits.circuits for op in sublist
-#         ]
-
-#         # Retrieve the corresponding solution data
-#         solution_data = s.experiment_circuit_solutions[solution_index]
-
-#         # Validate the operations
-#         for test_ops, solution_ops in zip(test_data, solution_data):
-#             for test_op, solution_op in zip(test_ops, solution_ops):
-#                 assert test_op.operation.name == solution_op.operation.name, (
-#                     f"Operation mismatch: {test_op.operation.name} != "
-#                     f"{solution_op.operation.name}"
-#                 )
-
+            assert len(subcirc.data) == sq.subcircuit_len[solution_index][circ_index]
 
 def test_expectation_values() -> None:
     """Test the expectation values of the test circuits.